sysv-generator: fix dependencies generated for services in runlevels [06]
If a SysV service is required to be started during shutdown (by
installing a 'S' symlink in /etc/init.d/rc0.d/) then the service
wasn't started at all because it was generated with
"Conflict=shutdown.target".

This patch fixes this case and also handles the weird case where both
'K' and 'S' symlinks are installed for the same service: the service
won't be started nor stopped at all. IOW it will be ignored during the
boot and shutdown process.