blkid: do not apply partition-type flags to ESP
Commit 0238d4c660e732dd03ba0cdb54a29ec5870ee849 said:

Apply them only to the types where they are defined, and not to the ESP,
which does not appear to share the same set of flags.

While fixing the issue in gpt-auto-generator, it introduced a new one in
builtin-blkid by doing exactly the contrary, applying them to the ESP.

This would cause blkid to fail applying the ID_PART_GPT_AUTO_ROOT_UUID tag
in some cases, leading to the /dev/gpt-auto-root symlink not being created.

fixes #3440