Correct default of `[Network] Private=` option for nspawn files
When manually starting a container using
`systemd-nspawn -bxD /var/lib/machine/foo`, it defaults to a non-private
network where the container gets the hosts interfaces.

When starting a container using `systemctl start systemd-nspawn@foo`
based on a config in `/etc/systemd/nspawn/foo.nspawn` (not containing a
`[Network]` section) it defaults to using `--network-veth` for
`systemd-nspawn`.