manager: Only invoke a single sigchld per unit within a cleanup cycle.
By default, each iteration of manager_dispatch_sigchld() results in a unit level
sigchld event being invoked. For scope units, this results in a scope_sigchld_event()
which can seemingly stall for workloads that have a large number of PIDs within the
scope. The stall exhibits itself as a SIG_0 being initiated for each u->pids entry
as a result of pid_is_unwaited().

This patch resolves this condition by only paying to cost of a sigchld in the underlying
scope unit once per sigchld iteration. A Set is instantiated at the
manager_dispatch_sigchld() entry, and is freed at the end of the iteration.