install: fix disable via unit file path
unit_file_disable currently chokes on full path. This is because
unit_name_is_valid returns -EINVAL if it receives full path. Arguably,
this is a correct behavior. Let's normalize a given string, take the
basename and work with that. If install_info_add would have been called
with path instead of name it would do the same anyway, so there is no
point in making distinction here.

Before:
        $ systemctl disable /usr/lib/systemd/system/avahi-daemon.service
        Failed to disable unit: Invalid argument

After:
        $ systemctl disable /usr/lib/systemd/system/avahi-daemon.service
        Removed /etc/systemd/system/dbus-org.freedesktop.Avahi.service.
        Removed /etc/systemd/system/sockets.target.wants/avahi-daemon.socket.
        Removed /etc/systemd/system/multi-user.target.wants/avahi-daemon.service.