man: mention system-shutdown hook directory in synopsis
The distinction between systemd-shutdown the binary vs system-shutdown
the hook directory (without the 'd') is not immediately obvious and can
be quite confusing if you are looking for a directory which doesn't
exist.

Thererfor explicitly mention the hook directory in the synopsis with a
trailing slash to make it clearer which is which.